refactor(api): split jellyfinApi into domain modules behind a barrel
services/jellyfinApi.ts was 4,326 lines holding every Jellyfin concern. It is
now a 74-line re-export barrel over 18 modules in services/jellyfin/: session,
connection, auth, demo, library, items, facets, search, userData, playback,
streamUrls, images, media, subtitles, http, constants, events, cacheKeys.

Public API unchanged: all 84 exports identical, every consumer still imports
from @/services/jellyfinApi. The barrel is load-bearing, not cosmetic. A dozen
test files mock the module by specifier, so a consumer importing a submodule
directly would slip past its own test's mock and hit the network while still
reporting green.

Placement notes:
- signOut lives in session.ts, not auth.ts. The 401 handler calls it, and auth
  already depends on session.
- library.ts keeps browse and view-root filter resolution together;
  isLibraryViewRoot calls fetchUserViews, so splitting them needs a cycle.
- cachedConfig is read through getCachedConfig() instead of a reassigned
  module binding.

Also extracts the request-timeout block into services/jellyfin/http.ts. 30
hand-rolled AbortController + setTimeout + double clearTimeout blocks collapse
to one fetchWithTimeout that clears the timer in a finally. Call sites keep
their own !ok branches and catch semantics, so no error text and no request
shape changed. That matters because retryWithBackoff classifies retryability
by regex-matching error.message. checkServerInfo is excluded: it composes an
external AbortSignal with paired listener cleanup for the LAN sweep.

Adds a dev-only playback regression probe (services/playbackProbe.ts,
scripts/playback-regression.mjs, test/playback baselines) wired through
player.tsx and useVideoPlayback behind a probe deep-link param.

44 suites / 834 tests green, tsc and eslint clean, no circular dependencies.
