+++ .gitignore
@@ -32,6 +32,7 @@ yarn-error.*
 
 # local env files
 .env*.local
+.env.playback-test
 
 # typescript
 *.tsbuildinfo
